Remove fixcolnames option added in #751
This was discussed and intentionally left out for all IO methods.

The plan is for `df.colname` to be the idiomatic way of specifying
a column in the near future (it already works on some experimental
AbstractDataFrame types), and `df.col.name`, for example, can't be
parsed as desired. (`.` is meaningful syntax in Julia, so using
`col.name` in place of a valid identifier is like using `col+name`
which wouldn't be valid in R, for example.

It's trivial to work around in user code if the user insists, but
it doesn't belong in any package code for now, though adding it
uniformly to all IO methods may be revisited later if the roadmap
changes.