Import R data frame attributes as metadata
Use the `metadata` function recently added to DataAPI and DataFrames
to import R `data.frame` attributes and set them as `DataFrame` metadata.
R stores per-column attributes in vector objects, while DataFrames.jl
stores them in the `DataFrame` object, as there is no generic mechanism
to attach metadata to an `AbstractVector` object.

The `row.names` attribute is skipped as it is not appropriate to store
it as global metadata given that it will get out of sync after subsetting
rows. We could provide a way to turn row names into a column instead.

Also add methods to check equality between two `DictoVec` objects
as these are useful for tests (haven commonly sets named numeric vectors
to store value labels so this case deserves testing).